DragDrop2 fixes
Having access to the browser event that triggered the Drop custom event
is useful (i.e. determining if CTRL/SHIFT/ALT key is pressed). Add
abilitiy to access this event object in custom event handlers.

DragDrop2:drop was passing the drop element as the memo attribute. This
should instead be the Drag object to be consistent with the over/out
events.

Improve documentation.