Further flesh out base Horde initialization.
There is two ways Horde applications can be initialized: through a
Horde_Registry API call and by directly loading the application (i.e.
loading a script in an application).  Both methods (may) require base
initialization of the application (things like setting up paths to the
template files, creating app-specific objects) but the latter also
requires that the app initially be pushed on the stack. Since this
latter behavior needs to be done in all applications, and is essentially
the same in all apps, move it to the globally accessible function
Horde_Registry::appInit().

Other benefits: initiation of compression can now be done in a single
place

Test out this new framework by converting Turba.
Add new syntax to Skeleton.