Consistenly allow setting prefix for cache entries.
For drivers where this makes sense, always allow setting a cache entry prefix
for preventing entry collisions from multiple horde instances on the same
physical box and for forcing the use of a certain prefix, regardless of the
servername we are hitting from. (This last point fixes an issue where a cache entry would
not expire when the entry was cached while accessing horde's data via the api from
a different virtual server).