Removed 'shared' feature of Horde_Prefs.
Defeats the whole purpose of decoupling the various applications.
Additionally, doesn't make much sense to declare horde
prefs 'shared' since they are always used as a fallback if the prefname
doesn't exist in the local scope.

Preference values from other applications should be accessed via an
application's API instead.  This is already what we do, for example, to
get the search_sources pref in IMP/Turba.

Not really sure how this was supposed to work anyway, since shared prefs
were totally dependent on the order in which the app was loaded. Simply
put, global prefs need to live in horde.