Prefs UI fixes
Remove prefsEnum() call; replace with prefsGroup() call

prefsInit() is designed for tasks that need to run once on the page.
This is where all suppressGroups handling should go.

Next, prefsGroups() is used to determine the list of prefs that need to
be updated.  However, after prefs are updated, suppressed prefs may
change so we need to run prefsGroups() again to get the updated list of
suppressed prefs for display.

prefsEnum() becomes redundant because it can be handled in either
prefsInit() or prefsGroups().