Completely rework feature hover handling
Use Symbolizers and StyleMaps on the Layer to implement custom
icon markers instead of passing a StyleHash for each marker. Besides
being more efficient, it's needed in order to place multiple markers
with differing style properties on the same layer. (If a feature has
it's own explicit StyleHash, it seems to override the Layer's StyleMap,
even for a different intent).