Play with adding a preview function for linked images.
For images returned in the tweet's media properties, add a preview icon
that opens an overlay and displays a small preview of the image. This is
first draft, still needs CSS love.

Hope to be able to expand this to pick up images from typical image
sharing links, and not limited to images uploaded directly via Twitter's API.