Work around chicken/egg problem with RPC auth.
RPC requests have authentication delegated to the RPC classes, so
we must bootstrap horde with authentication => none. Horde must be
initialized in horde to instantiate the objects needed to build the
Rpc server. However, once the Rpc object is authenticated to horde,
we need to tell $registry that we really *do* want authentication,
otherwise none of the authAuthenticateCallback() methods will ever
be called in any application.