Partially fix issue with sidebar generation due to removing ajaxui parameter.
This fixes non-ajax apps showing the sidebar while logged in as VIEW_DYNAMIC.
The problem still exists for apps that allow toggling the view preference in
the prefs (currently Kronolith and Hermes). If logged in as VIEW_DYNAMIC, but
have choosen VIEW_BASIC the sidebar will still not be shown.