Finish new button design.
The .button class is no longer necessary, all INPUTs of type "submit" and
"button" are styled as buttons now.

The following button classes exist and can applied to both INPUT and A tags:
- horde-button (only necessary for A tags)
- horde-default: for the default action button of a form.
- horde-delete: for desctructive buttons.
- horde-create: for "new/create" buttons.
- horde-cancel: for "cancel/close" buttons.

If possible, buttons should be wrapped in horde-form-buttons DIVs. This is the
default for Horde_Form button bars.