Return list of NoSQL drivers for the base Horde application
Dealing with MongoDB drivers this way is a bit hackish.  But we can't
take a "migration" approach like SQL databases for various reasons (not
to mention that the migration approach - at least the CLI method - is
pretty much worthless for anything other than the base SQL config set in
conf.php).

But this approach works and is clean from a UI perspective, so it
can/should be used for Horde 5.1.