Beginnings of a new dev install script.
THIS DOES NOT WORK YET.

It *does* correctly link the applications, set up the various include paths,
and do autoloading of framework libraries without having to install
packages with PEAR. (Yay)

The problem is the javascript files that live in the packages (actually, it
would be anything that is installed via the "horde" role) aren't
web-accessibly linked and thus things don't work very well. Not really sure
what to do about this.