Replace Kronolith-specific caldav_url hook with global davusername hook.
It's much simpler, sufficient for the most use cases, and has the big advantage that we can translate the user names back too.

The other DAV APIs beside Kronolith's CalDAV API still need to be updated to use this hook.

I wonder if we should generalize it even further to a "rpcusername" hook, because any RPC request suffers from the problem that users cannot select an authentication backend (IMAP server), which is what this hook is trying to solve.